requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Document Processing
PyPDF2>=3.0.0
ebooklib>=0.18
BeautifulSoup4>=4.12.0
pytesseract>=0.3.10

# NLP & AI
openai>=1.0.0
google-generativeai>=0.8.3

# Audio & Media
edge-tts>=6.1.0

# Image Processing
pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0


